Split redux reducers for chat and authentication
