+++ ahoy_frontend/src/components/Chat/FriendsList.js
@@ -57,7 +57,7 @@ class FriendsList extends React.Component {
   );
 }
 
-const mapStateToProps = (state) => {
-  return { userInfo: state.userInfo };
+const mapStateToProps = ({ authentication }) => {
+  return { userInfo: authentication.userInfo };
 };
 export default connect(mapStateToProps, { openUserChat })(FriendsList);
