+++ ahoy_frontend/src/components/Chat/OpenChats.js
@@ -37,7 +37,8 @@ class OpenChats extends React.Component {
   );
 }
 
-const mapStateToProps = ({ openChats, userInfo }) => {
-  return { openChats: openChats, userInfo: userInfo };
+const mapStateToProps = ({ authentication, chat }) => {
+  console.log(authentication);
+  return { openChats: chat.openChats, userInfo: authentication.userInfo };
 };
 export default connect(mapStateToProps, { receiveMessage })(OpenChats);
