+++ ahoy_frontend/src/components/TopBar.js
@@ -41,7 +41,7 @@ const TopBar = (props) => (
   </Navbar>
 );
 
-const mapStateToProps = (state) => {
-  return { userInfo: state.userInfo };
+const mapStateToProps = ({ authentication }) => {
+  return { userInfo: authentication.userInfo };
 };
 export default connect(mapStateToProps)(TopBar);
