+++ ahoy_frontend/src/components/UserProfile.js
@@ -61,7 +61,7 @@ class UserProfile extends React.Component {
   );
 }
 
-const mapStateToProps = (state) => {
-  return { userInfo: state.userInfo };
+const mapStateToProps = ({ authentication }) => {
+  return { userInfo: authentication.userInfo };
 };
 export default connect(mapStateToProps)(UserProfile);
